Fix bug where file with only removed lines was being detected as new
Checker.linesChanged returns a map of files and a slice of positions
that have changed. If the slice was nil, it was an indication that
the file was new, so any issues found in that file were new and should
be raised. But if the file only had lines that were removed (not added
or modified), the slice would also be new.

This change introduces a third state, where an empty slice would indicate
no lines have been added (or modified).

I don't think this is an ideal change, the bug was likely introduced
in 8d3df443d0743c5b4346687e60ae485d75fbe3be, where previous to that
new files had an artificial diff created indicating all lines were
added. That commit was designed to remove the external dependency in
GNU diff.

I've settled on this change simply because it's straight forward and
doesn't add another dependency, but I'm not convinced it's the best
change either.

There's other related issues to this, some tools may want to see issues
on lines that have been removed, such as apicompat, a larger change
may want to take this into account.

Related to https://github.com/bradleyfalzon/gopherci/issues/86.
Related to https://github.com/bradleyfalzon/gopherci/issues/63.